+++ injection_scripts/wendigo_utils.js
@@ -3,7 +3,7 @@
 window.WendigoUtils = {
     isVisible(element) {
         if(!element) return false;
-        let style = window.getComputedStyle(element, "");
+        const style = window.getComputedStyle(element);
         if (style.display === 'none') return false;
         if (style.visibility === 'hidden') return false;
         else return true;
@@ -16,9 +16,12 @@ window.WendigoUtils = {
     queryAll(selector) {
         if(typeof selector === 'string') {
             return document.querySelectorAll(selector);
-        } else if(!Array.isArray(selector)) {
-            return [selector];
-        } else return selector;
+        } else{
+            if(!Array.isArray(selector)) {
+                selector = [selector];
+            }
+            return selector;
+        }
     },
     xPathQuery: function(xPath) {
         let xPathResult = document.evaluate(xPath, document, null, XPathResult.ANY_TYPE, null);
@@ -27,6 +30,15 @@ window.WendigoUtils = {
         while(r !== null) {
             result.push(r);
             r = xPathResult.iterateNext();
+        }
+        return result;
+    },
+    getStyles(element, pseudoSelector) {
+        const rawStyles = getComputedStyle(element, pseudoSelector);
+        const result = {};
+        for(let i = 0;i < rawStyles.length;i++) {
+            const name = rawStyles[i];
+            result[name] = rawStyles.getPropertyValue(name);
         }
         return result;
     }